photo of masks
this is the masks using four sheets of fabrics from a 100% cotton T-shirt, using a 7" elastic band.  There are issues of the pleats at both sides are too thick when folding 4x3 = 12 sheets together and take too long to sew them together.  The needle is bended.    Unfortunately I have not find any alternative yet.   The other issue is the ear loop using 7" elastic band (with effective length 6") is too short and hurt the ear.  Adding length to 8 to 8.5"  with effective length of 7"-7.5"  helps.  This illustrate the need for adjustment by user if so desired.